cambios

+++ Codere/Models/Show.cs
@@ -22,7 +22,7 @@ namespace Codere.Models
          public string? OfficialSite { get; set; }
          public int? Weight { get; set; }
          public string? Summary { get; set; }
-         public Double? Updated { get; set; }
+         public long? Updated { get; set; }
          public double? Rating_Average { get; set; }
